Move `iou2d_calculator` to utils
