+++ src/otx/algo/detection/utils/assigner/max_iou_assigner.py
@@ -1,5 +1,6 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
+# Copyright (c) OpenMMLab. All rights reserved.
 """Max Iou Assigner implementation from mmdet."""
 
 from __future__ import annotations
@@ -10,7 +11,7 @@ from typing import TYPE_CHECKING, Callable
 import torch
 from torch import Tensor
 
-from otx.algo.detection.heads.iou2d_calculator import BboxOverlaps2D
+from otx.algo.detection.utils.iou2d_calculator import BboxOverlaps2D
 from otx.algo.detection.utils.structures import AssignResult
 
 if TYPE_CHECKING:
