+++ src/otx/algo/detection/utils/assigner/sim_ota_assigner.py
@@ -11,7 +11,7 @@ import torch
 import torch.nn.functional as F  # noqa: N812
 from torch import Tensor
 
-from otx.algo.detection.heads.iou2d_calculator import BboxOverlaps2D
+from otx.algo.detection.utils.iou2d_calculator import BboxOverlaps2D
 from otx.algo.detection.utils.structures import AssignResult
 
 if TYPE_CHECKING:
