+++ src/otx/algo/detection/utils/iou2d_calculator.py
@@ -1,5 +1,6 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
+# Copyright (c) OpenMMLab. All rights reserved.
 """2D IoU calculator implementation from mmdet."""
 
 from __future__ import annotations
